Move schemas folder
